Add files via upload
Linear regression model to predict house prices in a city. The model is built by feeding it simple data created on Excel and converted to a CSV file and read with Pandas. Linear Regression model gotten through Linear_Model from Sklearn (Scikit-Learn) library was used to fit the dataset and to make prediction. Finally, the result was plotted using Matplotlib.pyplot on a scatter plot to display the predicted values.